update etcd to consul for k/v
This commit updates all k/v operations
from etcd to consul.

NOTE - the way that k/v lookup works in consul
is a little different and may not scale in the
same way as the etcd version.

NOTE - this is not yet completed, it is just being
published for early review.