Don't upgrade on 200/404 status codes
Currently, a 500 returned form the consul server
(in our discovered example, when its leader is
swapping due to network unreliability) results
in all machines receiving an upgrade signal, b/c
the upgrade system is based on default to yes.

This patch Adds code to ensure that upgrades only
occur in cases where either a result is returns that
is set to false, or if there are no keys (ie
status codes 200 and 404). All other return codes
are treated explicitly as do-not-upgrade.